BreastCancerDetection/model_service/app.py
from flask import Flask, request, jsonify
import numpy as np
import base64
import io
from PIL import Image
import time
import random

app = Flask(__name__)

# Simulated model prediction function
def analyze_image(image_data):
    """
    Simulated function to analyze breast ultrasound images.
    In a real application, this would use a trained CNN-LSTM model.
    """
    # Simulate processing time
    time.sleep(1)
    
    # Random prediction (benign or malignant) with confidence
    prediction = "benign" if random.random() > 0.3 else "malignant"
    confidence = random.uniform(0.7, 0.99)
    
    # Generate simulated feature values
    features = {
        "radius": random.uniform(6.5, 28.1),
        "texture": random.uniform(9.7, 39.3),
        "perimeter": random.uniform(43.8, 188.5),
        "area": random.uniform(143.5, 2501.0),
        "smoothness": random.uniform(0.05, 0.16),
        "compactness": random.uniform(0.02, 0.35),
        "concavity": random.uniform(0.0, 0.43),
        "concave_points": random.uniform(0.0, 0.2),
        "symmetry": random.uniform(0.1, 0.3),
        "fractal_dimension": random.uniform(0.04, 0.1)
    }
    
    return {
        "prediction": prediction,
        "confidence": confidence,
        "features": features
    }

@app.route('/predict', methods=['POST'])
def predict():
    if not request.json or 'image' not in request.json:
        return jsonify({'error': 'No image data provided'}), 400
        
    try:
        # Get base64 image data
        image_data = request.json['image']
        
        # In a real application, decode and process the image
        # image_bytes = base64.b64decode(image_data)
        # image = Image.open(io.BytesIO(image_bytes))
        
        # Analyze image
        result = analyze_image(image_data)
        
        return jsonify(result), 200
        
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({'status': 'healthy', 'model': 'breast-cancer-detection-v1'}), 200

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)
