BreastCancerDetection/model_service/requirements.txt
Flask==2.3.2
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
cors==1.0.1
